Refactor filterTodo to it's own file and test it
